requirements.txt
streamlit==1.53.0
numpy==2.2.3
pillow==11.1.0
faiss-cpu==1.10.0
qdrant-client==1.16.0
python-dotenv==1.0.1
pydantic==2.10.6
PyYAML==6.0.2
tqdm==4.67.1
opencv-python==4.12.0.88
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
pandas==2.3.3